implement conventional syntax for functions
