+++ src/main/scala/io/chymyst/ch/TheoremProver.scala
@@ -24,7 +24,7 @@ object TheoremProver {
     val proofTerms = findProofTerms(mainSequent).map(_.prettyRename).distinct
     if (debug) {
       val prettyPT = proofTerms.map(p ⇒ (p.prettyPrint, p.unusedArgs.size, p.unusedTupleParts, p.unusedArgs, p.usedTuplePartsSeq.distinct.map { case (te, i) ⇒ (te.prettyPrint, i) }))
-        .sortBy { case (pString, s1, s2, unusedArgs, usedTupleParts) ⇒ s1 + s2 }
+        .sortBy { case (_, s1, s2, _, _) ⇒ s1 + s2 }
       println(s"debug: got proof terms:\n ${prettyPT.mkString(";\n ")}")
     }
     // Return the group of proofs that leave the smallest number of values unused, but has the smallest use count of those that are used.
